Significant cleanup in replay recorder, updater, refactoring of path, etc.
