+++ src/fa/play.py
@@ -1,3 +1,22 @@
+# -------------------------------------------------------------------------------
+# Copyright (c) 2012 Gael Honorez.
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the GNU Public License v3.0
+# which accompanies this distribution, and is available at
+# http://www.gnu.org/licenses/gpl.html
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#-------------------------------------------------------------------------------
+
+
 from process import instance
 
 import util
@@ -11,6 +30,8 @@ from PyQt4 import QtCore
 
 settings = QtCore.QSettings("ForgedAllianceForever", "FA Lobby")
 
+from . import DEFAULT_WRITE_GAME_LOG
+from . import DEFAULT_RECORD_REPLAY
 
 def build_argument_list(game_info, port, arguments=None):
     """
@@ -25,12 +46,12 @@ def build_argument_list(game_info, port, arguments=None):
         arguments.append("init_" + game_info['featured_mod'] + ".lua")
 
     #log file
-    if settings.value("fa.play.write_game_log", type=bool):
+    if settings.value("fa.write_game_log", DEFAULT_WRITE_GAME_LOG, type=bool):
         arguments.append("/log")
         arguments.append('"' + util.LOG_FILE_GAME + '"')
 
     #live replay
-    if settings.value("fa.play.stream_live_replay", type=bool):
+    if settings.value("fa.record_replay", DEFAULT_RECORD_REPLAY, type=bool):
         arguments.append('/savereplay')
         arguments.append('gpgnet://localhost/' + str(game_info['uid']) + "/" + str(game_info['recorder']) + '.SCFAreplay')
